Implements cache files to store covid stats for specific locations

+++ app/src/main/java/app/covidstats/model/data/covid_stats/CovidStats.kt
@@ -7,7 +7,7 @@ data class CovidStats(
     val active: Int,
     val activePerOneMillion: Double,
     val cases: Int,
-    val casesPerOneMillion: Int,
+    val casesPerOneMillion: Double,
     val critical: Int,
     val criticalPerOneMillion: Double,
     val deaths: Int,
